pyproject.toml
[project]
name = "nvidia"
version = "0.1.0"
description = "Backend para sistema de segurança com agentes de IA da NVIDIA"
authors = [
    {name = "FaraoBack",email = "cairebelok3@gmail.com"}
]
readme = "README.md"
requires-python = "3.11.2"
dependencies = [
    "crewai (>=0.157.0,<0.158.0)",
    "crewai-tools (>=0.60.0,<0.61.0)",
    "openai (>=1.99.1,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "llama-index (>=0.13.0,<0.14.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pika (>=1.3.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
